app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pillow==11.0.0
ultralytics==8.3.40
torch==2.5.1
torchvision==0.20.1
clip @ git+https://github.com/openai/CLIP.git
